Add files via upload
This was made by watching Mosh's tutorial on YouTube! This utilizes Express for CRUD operations and Joi for validation of the content (character length and requiring a "name" key with an attached value with a character length of over 3 characters). 
This has been a most enlightening tutorial for me and great practice with Express functionality and ES6 syntax.